requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==0.16
Flask-WTF==0.9.5
Flask-WhooshAlchemy==0.55a
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
SQLAlchemy==0.7.9
Tempita==0.5.2
WTForms==1.0.5
Werkzeug==0.9.6
Whoosh==2.6.0
argparse==1.2.1
blinker==1.3
decorator==3.4.0
gunicorn==19.0.0
itsdangerous==0.24
nltk==2.0.4
pysqlite==2.6.3
sqlalchemy-migrate==0.7.2
wsgiref==0.1.2
